Separate error handling from core functionality
To de-clutter the code further, we can introduce a function containing
the code that we want to run, and invoke that function from `main()`
with all the error handling done there. This keeps the important code
free of error handling, and easier to read and maintain.

We also take the opportunity to pass the arguments to our new function
in the expected format. This leaves a skeleton main() function that we
can leave as-is, and we can now focus on the core functionality.