Add function to compute overlap between sequence and candidate fragment
To compute the overlap, we need to first agree on how to represent
overlap. Here, we define overlap as an integer, corresponding to the
number of characters at the start of the main sequence that constitute
the overlap region. This implies that the overlapping part on the
candidate fragment is the same number of characters at the *end* of the
fragment.

We then search for overlap by comparing these two sub-strings, starting
from the largest possible overlap: the size of the full fragment. The
reason is that we want to check for the largest possible overlap.
Starting from the smallest overlap and increasing the size means the
first match encountered may not be the largest possible match. Starting
from the largest means we can report the value of the overlap as soon as
we have a match, and break out of the loop at that point.