Move overlap functionality into ShotgunSequencer class
As the functionality provided in the overlap.h header really only made
sense when used as part of the solver, it makes sense to include these
functions and the Overlap struct as part of the ShotgunSequencer class.
Here, we make them private methods, and declare the Overlap struct
within the ShotgunSequencer class. This allows us to use these methods
internally, but makes them inaccessible from outside the class,
effectively hiding the internals of the algorithm - an example of
abstraction.