Introduce debug::log() convenience function
Since it's likely we'll want to add lots of debugging statements, it's
helpful to define a small function to reduce the amount of code
duplication and improve readability. In this case, we define an inline
function as this is a short function. Furthermore, using an inline
function allows the compiler to optimise away the function call, which
can help with performance.