Add another option to specify the pixel of interest
We add the "-p" option to allow the user to specify which voxel to
inspect. This requires taking out 3 command-line arguments: -p x y.
There are several ways to do this, but arguably the cleanest way relies
on STL iterators. Since many STL algorithms rely on iterators, it is
worth trying to understand them.