Introduce plotting on the terminal
We have provided a small library that you can use to plot on the
terminal. This uses an ancient technology called sixel graphics
(originally designed for dot matrix printers!), which can still be used
with the right terminal. The terminal we are using on the course (mintty
with a MSYS2 MSYS session) happens to support this straight away. On
macOS, the best option is to install iTerm2 and use that instead of the
standard terminal.