Introduce operator overloading for Image & Dataset classes
We can use operator overloading to provide a more intuitive syntax to
access pixel values (for the Image class), or access individual data
slice / time-points (for the Dataset class).

This is also useful since this is exactly the syntax that the
terminal_graphics library expects for images. We can now display the
first time point by passing it to the TG::imshow() call. Since the image
is otherwise too small to see clearly, we also use the TG::magnify()
functionality to make it bigger.