Introduce lambda functions for task rescaling
To display the task alongside the signal timecourse, we need to rescale
the task to match the overall intensity and range of the signal.
The operation can be performed in different ways, but perhaps the most
useful and compact is the lambda expression, which we use here.

It also allows us to declare the operation at the point of use, which
for such a simple operation, is perhaps more expressive than hiding it
away in a different file.

This is also where STL algorithms can be useful: they provide ways of
applying basic operations to containers, where the specific operation to
be applied can be defined using lambda functions (amongst other
possibilities).