Add our first concrete Segment: the tip
It is not currently possible to create an instance for the Base class,
since we need to provide an instance of another segment in the
constructor. But we can define a new segment for the tip: Segment::Tip,
and in its constructor, simply provide a reference to itself. Note that
this reference could be set to anything, since the tip will never
attempt to access the next segment - but since the current instance
exists and is of a valid type, we may as well use it here. Besides, it
makes sense that the next segment after the tip is still the tip...

Now that we can create the tip segment, we can also create an instance
of the base segment.