Load parameters from file for full robot arm configuration
We can now add functionality to load the parameters for robot arm from
the files provided. To do this correctly, we introduce the stringstream
class, which allows us to read a full line from the file, and then read
from that line as if it was an independent stream. This allows us to
get the full set of parameters from each line without knowing how many
parameters should be present on each line, and then check that all
lines contain the same number of parameters.

Finally, we can assign the parameters we loaded into the robot arm, and
report the tip position for each time point.