Use a smart pointer to manage the memory
We can use a smart pointer to manage all this for us. The std::shared
pointer is suitable here, as it allows copies to be passed around, and
only delete the instance once all references to it have been destroyed.
This means we don't have to worry about the destructor, copy constructor
or assignment operator, and just leave these as default values.

We can then replace the use of the new operator with the
std::make_shared() template call, which will create a shared pointer of
the requested type, and forward the arguments provided to the class
constructor.

An alternative would be to use the std::unique_ptr smart pointer, but it
does not provide a copy constructor or assignment operator (these have
been deleted) - as expected if it is to be the sole owner. It could be
used, but this would require the use of move semantics, which are beyond
the scope of this course.